requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyGithub
httpx